Move `CleverLibrary::Api::Oauth` to `CleverIntegration::OauthClient`
